+++ packages/import/client/views/panels.js
@@ -14,21 +14,50 @@ Template.formImport.helpers({
 	},
 	importOnSelect: function() {
 		var tmpl = Template.instance();
-		return function(target, fileObj, params) {
-			var name$ = tmpl.$('.import-name');
-			name$.val( K.Import.importnameFromFile(fileObj) );
+		return function(err, fileObj, params) {
+			var name$ = tmpl.$('.import-name'),
+				importName =  K.Util.sanitize.importName(fileObj.name);
+			
+			var geoj = JSON.parse(fileObj.blob);
+
+			var n=0, maxFeatures = 10,
+				sample = _.first(_.shuffle(geoj.features), maxFeatures);
+
+			K.Map.addGeojson({features: sample});
+
+			_.each(sample, function(f) {
+				
+				let cen = K.Util.geo.centroid(f.geometry),
+					icon = new L.NodeIcon(),
+					marker = L.marker(cen, {icon: icon});
+
+				marker.addTo(K.Map.layers.geojson);
+				
+				Blaze.renderWithData(Template.markerPlace, f.properties, icon.nodeHtml);
+				n++;
+			});
+
+			K.Alert.info(i18n('label_importpreview',n,geoj.features.length));
+
+			name$.val( importName );
 		}
 	},
 	importOnUploaded: function() {
 		var tmpl = Template.instance();
-		return function(ret) {
-			K.Alert.info(ret)
+		return function(ret, fileObj, params) {
+			K.Map.cleanGeojson();
+			K.Alert.info(ret);
 		}
 	}
 });
 
 Template.formImport.events({
-	'change .import-name': function(e, tmpl) {
-		tmpl.importName.set(e.target.value)
-	}
+	'keydown .import-name': _.debounce(function(e, tmpl) {
+		console.log(e.target.value)
+		var importName = K.Util.sanitize.importName(e.target.value);
+		
+		$(e.target).val(importName);
+
+		tmpl.importName.set(importName)
+	},300)
 });