+++ packages/import/i18n/en.js
@@ -3,6 +3,7 @@ i18n.map('en', {
 	label_import: "Import Data",
 	label_importfile: "Upload a GeoJSON File",
 	label_importname: "Import Name",
+	label_importpreview: "Import Preview ({$1} places of {$2} totals)",
 	label_imported: "Imported places",
 	label_imports: "Imports",
 	error_import_formatNotValid: "Format not Valid",