+++ packages/import/modules/Import.js
@@ -1,6 +1,0 @@
-
-Kepler.Import = {
-	importnameFromFile: function(fileObj) {
-		return K.Util.sanitize.fileName(fileObj.name, true) || K.Util.timeName();
-	}
-};