+++ packages/import/server/Import.js
@@ -3,25 +3,21 @@ Kepler.extend({
 	Import: {
 		geojsonToPlace: function(feature, importName) {
 
-			var prop = feature.properties,
-				coords = feature.geometry.coordinates,
-				loc = [coords[1], coords[0]];
-
-			var name = prop.name || '';
-			var url = prop.url || '';
-
-			_.each(prop, function(v, k) {
-				if(!v) {
-					delete prop[k];
-				}
-			});
+			var coords = feature.geometry.coordinates,
+				props = feature.properties,
+				loc = [];
 
 			if(feature.geometry.type==='Point') {
+				loc = [coords[1], coords[0]];
+			}
+			//TODO else centroid
+
+			if(K.Util.valid.loc(loc)) {
+
 				return {
-					name: K.Util.sanitize.name(name),
-					url: K.Util.sanitize.url(url),
-					loc: K.Util.geo.locRound(loc,8),
-					active: 0,
+					name: K.Util.sanitize.name(props.name || ''),
+					url: K.Util.sanitize.url(props.url || ''),
+					loc: K.Util.geo.locRound(loc, 8),
 					import: {
 						name: importName,
 						data: feature