+++ packages/upload/client/Upload.js
@@ -1,11 +1,49 @@
 
 Kepler.Upload = {
+
+	fileReader: null,
 	
+	loadFile: function(target, fileObj, params, callback) {
+
+		callback = _.isFunction(callback) ? callback : function(){};
+
+		var sets = K.settings.public.upload.targets[target],
+			err = null
+		
+		sets.maxFileSize = sets.maxFileSize || K.settings.public.upload.maxFileSize
+
+		if(!fileObj) return false;
+
+		if(fileObj.size > sets.maxFileSize) {
+			err = i18n('upload_error_filesizeNotValid') + 
+				  K.Util.humanize.filesize(sets.maxFileSize);
+			callback(err);
+			return this;
+		}
+
+		if(this.fileReader)
+			this.fileReader.abort();
+		else
+			this.fileReader = new FileReader();
+
+		this.fileReader.onloadend = function(e) {
+			fileObj = {
+				name: fileObj.name,
+				type: fileObj.type,
+				size: fileObj.size,
+				blob: e.target.result
+			};
+			callback(err, fileObj, params)
+			//Meteor.call('uploadFile', target, fileObj, params, callback);
+		};
+		this.fileReader.readAsBinaryString(fileObj);
+	},
+
 	uploadFile: function(target, fileObj, params, callback) {
 
 		callback = _.isFunction(callback) ? callback : function(){};
 
-		var sets = K.settings.public.upload.targets[target];
+		/*var sets = K.settings.public.upload.targets[target];
 		
 		sets.maxFileSize = sets.maxFileSize || K.settings.public.upload.maxFileSize
 
@@ -15,7 +53,7 @@ Kepler.Upload = {
 			callback( i18n('upload_error_filesizeNotValid') + 
 				K.Util.humanize.filesize(sets.maxFileSize) );
 			return this;
-		}
+		}*/
 
 		if(_.isObject(sets.mimeFileType)) {
 			var mimes = [];
@@ -30,21 +68,11 @@ Kepler.Upload = {
 			}
 		}
 
-		if(this.fileReader)
-			this.fileReader.abort();
-		else
-			this.fileReader = new FileReader();
+		if(!this.fileReader)
+			this.loadFile(target, fileObj, params)
+
+		Meteor.call('uploadFile', target, fileObj, params, callback);
 		
-		this.fileReader.onloadend = function(e) {
-			fileObj = {
-				name: fileObj.name,
-				type: fileObj.type,
-				size: fileObj.size,
-				blob: e.target.result
-			};
-			Meteor.call('uploadFile', target, fileObj, params, callback);
-		};
-		this.fileReader.readAsBinaryString(fileObj);
 		//TODO not work.. this.fileReader.readAsDataURL(fileObj);
 		
 		return this;