+++ CHANGELOG.md
@@ -7,11 +7,16 @@
 - core rename methods K.Util.geo.bboxContains(), bboxReverse(), bboxPlain(), bboxRound() 
 - core rename methods K.Util.geo.locBuffer(), locTransform(), locRound()
 - core simplify method K.Util.geo.bboxContains()
-- core new method L.Util.geo.polygonCentroid()
+- core new method L.Util.geo.centroid() support Point, Polygon, MultiPolygon
 - plugin upload rename param 'callback' in 'onUploaded' in template inputFile_upload
+- plugin upload new param 'params' in template inputFile_upload
+- plugin upload new method K.Upload.loadFile()
 - plugin import K.Import.geojsonToPlace()
-- plugin import added new place holder panelImport
+- plugin import new params importname by form
+- plugin import new global method K.Util.sanitize.importName
 - core K.Util.geo.point() accept [lat,lon]
+- core edit outputof K.Util.timeName()
+- core fix K.Util.sanitize.fileName()
 
 ## v1.6.6
 - core support Instagram OAuth signup